refactor: add functions in separate files
